# openapi/v1.yml
...
openapi: "3.0.0"
info:
  version: "1.0.0-alpha1"
  description: "API definition for Kleister, manage mod packs for Minecraft"
  title: "Kleister OpenAPI"
  contact:
    name: "Kleister"
    email: "kleister@webhippie.de"
    url: "https://kleister.eu"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://try.kleister.eu/api/v1"
security:
  - Header: []
  - Bearer: []
  - Basic: []
  - Cookie: []
paths:
...
  /packs/{pack_id}/builds:
    get:
      summary: "Fetch all available builds for a pack"
      operationId: "ListBuilds"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
        - name: "sort"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "name"
            enum:
              - "name"
              - "public"
          description: "Sorting column"
          x-example: "name"
        - name: "order"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "asc"
            enum:
              - "asc"
              - "desc"
          description: "Sorting order"
          x-example: "desc"
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
          description: "Paging limit"
          x-example: 100
        - name: "offset"
          in: "query"
          required: false
          schema:
            type: "integer"
          description: "Paging offset"
          x-example: 0
      responses:
        "200":
          description: "A collection of builds"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/builds"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    post:
      summary: "Create a new build for a pack"
      operationId: "CreateBuild"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
      requestBody:
        description: "The build data to create"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/build"
      responses:
        "200":
          description: "The created build build"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/build"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
  /packs/{pack_id}/builds/{build_id}:
    get:
      summary: "Fetch a specific build for a pack"
      operationId: "ShowBuild"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
        - in: "path"
          name: "build_id"
          description: "A build identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "n3eked21ybxvfw2zi25u"
      responses:
        "200":
          description: "The fetched build details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/build"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Build or pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    put:
      summary: "Update a specific build for a pack"
      operationId: "UpdateBuild"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
        - in: "path"
          name: "build_id"
          description: "A build identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "n3eked21ybxvfw2zi25u"
      requestBody:
        description: "The build data to update"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/build"
      responses:
        "200":
          description: "The updated build details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/build"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Build or pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    delete:
      summary: "Delete a specific build for a pack"
      operationId: "DeleteBuild"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
        - in: "path"
          name: "build_id"
          description: "A build identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "n3eked21ybxvfw2zi25u"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Build or pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "400":
          description: "Failed to delete the build"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
...
components:
  securitySchemes:
    Header:
      type: apiKey
      in: header
      name: X-API-Key
    Bearer:
      type: http
      scheme: bearer
    Basic:
      type: http
      scheme: basic
    Cookie:
      type: apiKey
      in: header
      name: Cookie
  schemas:
...
    minecraft:
      title: "Minecraft"
      description: "Model to represent minecraft"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    forge:
      title: "Forge"
      description: "Model to represent forge"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        minecraft:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    neoforge:
      title: "Neoforge"
      description: "Model to represent neoforge"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    quilt:
      title: "Quilt"
      description: "Model to represent quilt"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    fabric:
      title: "Fabric"
      description: "Model to represent fabric"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    pack:
      title: "Pack"
      description: "Model to represent pack"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        icon:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_icon"
        logo:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_logo"
        back:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_back"
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        website:
          type: "string"
          x-omitempty: true
          x-nullable: true
        public:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    builds:
      title: "Builds"
      description: "Model to represent list of builds"
      type: "object"
      properties:
        total:
          type: integer
          format: int64
        pack:
          readOnly: true
          $ref: "#/components/schemas/pack"
        builds:
          type: "array"
          items:
            $ref: "#/components/schemas/build"
    build:
      title: "Build"
      description: "Model to represent build"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        pack:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack"
        minecraft_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        minecraft:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/minecraft"
        forge_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        forge:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/forge"
        neoforge_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        neoforge:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/neoforge"
        quilt_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        quilt:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/quilt"
        fabric_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        fabric:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/fabric"
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        java:
          type: "string"
          x-omitempty: true
          x-nullable: true
        memory:
          type: "string"
          x-omitempty: true
          x-nullable: true
        latest:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        recommended:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        public:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    notification:
      title: "Notification"
      description: "Generic response for errors and validations"
      type: "object"
      properties:
        status:
          type: "integer"
          x-omitempty: true
        message:
          type: "string"
          x-omitempty: true
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/validation"
    validation:
      title: "Validation"
      description: "General structure to show validation errors"
      type: "object"
      properties:
        field:
          type: "string"
        message:
          type: "string"
...